feat: 扩展文章更新请求模型支持编辑锁机制
- ArticleUpdateReq新增editToken和version字段，支持编辑权限验证
- 修复ArticleVO.currentVersion字段类型，统一使用Long类型
- 添加必要的validation注解，确保编辑操作数据完整性
- 为双版本编辑机制提供完整的数据传输支持

+++ bytelogs-api/src/main/java/top/harrylei/forum/api/model/vo/article/vo/ArticleVO.java
@@ -110,7 +110,7 @@ public class ArticleVO extends BaseVO {
      * 当前发布版本号
      */
     @Schema(description = "当前发布版本号", example = "1")
-    private Integer currentVersion;
+    private Long currentVersion;
 
     /**
      * 文章内容
